Displayed park titles that the user has completed and the ability to delete it from your account.
