+++ src/models/TrendingCardModel.ts
@@ -6,6 +6,7 @@ interface image {
 export default interface TrendingCardsModel {
   uid?: string;
   id: string;
+  _id?: string;
   images: image[];
   fullName: string;
   description: string;
